Rename 4-helpers to 8-helpers to make room for react-ecsal layer [Re: nmkataoka/adv-life#101]
