+++ src/1-game-code/Agent/ConditionSet/ConditionSet.ts
@@ -1,5 +1,5 @@
 import { EntityManager } from '0-engine';
-import { swapRemoveAt } from '4-helpers/ArrayExtensions';
+import { swapRemoveAt } from '8-helpers/ArrayExtensions';
 import { ComponentTemplateBase } from './ComponentTemplate';
 import { EntityTemplateBase } from './EntityTemplate';
 import { EntityRelationshipTemplateBase } from './EntityRelationshipTemplate';
