+++ src/5-react-components/Modal/index.test.tsx
@@ -1,15 +1,17 @@
 import React from 'react';
 import * as sinon from 'sinon';
-import {
-  render, fireEvent, screen, waitFor,
-} from '4-helpers/test-utils';
+import { render, fireEvent, screen, waitFor } from '8-helpers/test-utils';
 
 import Modal from '.';
 
 describe('<Modal />', () => {
   it('when close button is clicked, closes and calls onClose', async () => {
     const handleClose = sinon.stub();
-    render(<Modal isShowing onClose={handleClose}><div>Hello!</div></Modal>);
+    render(
+      <Modal isShowing onClose={handleClose}>
+        <div>Hello!</div>
+      </Modal>,
+    );
 
     fireEvent.click(screen.getByRole('button'));
