+++ src/6-ui-features/CombatScene/actions.ts
@@ -1,6 +1,6 @@
 import { createAction } from '@reduxjs/toolkit';
 import { getUnitInfos, UnitInfo } from '3-frontend-api/UnitInfo';
-import { DictOf } from '4-helpers/DictOf';
+import { DictOf } from '8-helpers/DictOf';
 import { AppThunk } from '7-app/types';
 
 export const updatedUnits = createAction<{ units: DictOf<UnitInfo> }, 'combatScene/updatedUnits'>(
