+++ src/6-ui-features/Items/itemClassesSlice.ts
@@ -1,7 +1,7 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { getItemClasses } from '3-frontend-api/Items/getItemClasses';
 import { ItemClassInfo } from '3-frontend-api/Items/ItemClassInfo';
-import { DictOf } from '4-helpers/DictOf';
+import { DictOf } from '8-helpers/DictOf';
 import { AppThunk } from '7-app/types';
 
 const initialState = {
