+++ src/6-ui-features/TownLocation/townLocationsSlice.ts
@@ -3,7 +3,7 @@ import deepEqual from 'fast-deep-equal';
 import { BUY_ITEM } from '2-backend-api/controllers/ShopConstants';
 import apiClient from '3-frontend-api/ApiClient';
 import { getTownLocationInfo, TownLocationInfo } from '3-frontend-api/town';
-import { DictOf } from '4-helpers/DictOf';
+import { DictOf } from '8-helpers/DictOf';
 import { AppThunk } from '7-app/types';
 import { playerBoughtItemFromShop } from '../Player/playerBoughtItemFromShop';
