+++ src/6-ui-features/Towns/townSlice.ts
@@ -1,7 +1,7 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { TownInfo } from '3-frontend-api/town';
 import { getTowns } from '3-frontend-api/town/getTowns';
-import { DictOf } from '4-helpers/DictOf';
+import { DictOf } from '8-helpers/DictOf';
 import { AppThunk } from '7-app/types';
 
 const initialState = {
