+++ src/8-helpers/DictOf.ts
